Rename baths module into environment
