+++ qutip/environment.py
